Refactor: Extract new mod `crate::utils`
