+++ crates/binstalk-downloader/src/lib.rs
@@ -2,3 +2,5 @@
 
 pub mod download;
 pub mod remote;
+
+mod utils;
